Shop implemented (#32)
* Fixed issue with login and register

* Removed error causing namespaces

* Added coin

* Coin display

* Added shop window and some design changes

* Added the possibility to double coins and buy new cars

* Made coin collection accumulate

* Added way to select the skin

* Fixed skin selection and coin debug deletion

* Added more player cars

* Fixed car selection bug

+++ Assets/Scripts/Client/DataTypes/ApiTypes/Game.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Text.Json.Serialization;
 
 namespace Thesis_backend.Data_Structures
 {
@@ -11,7 +10,7 @@ namespace Thesis_backend.Data_Structures
         public User User { get; set; }
         public long CurrentXP { get; set; }
         public int NextLVLXP { get; set; }
-        public long Currency { get; set; }
+        public int Currency { get; set; }
         public List<OwnedCar>? OwnedCars { get; set; }
 
         public override object Serialize => new { ID, Lvl, CurrentXP, NextLVLXP, UserId, Currency, OwnedCars };
